Lazily compute palette
This makes it so TipImage::Palette() is a lot faster after being called once.
Libtip/Tipdump code does the right thing and only ever calls it once,
but in edge cases (see: user code) this should improve perf a tiny bit.

I wonder if it would be possible to do some magic to make lazy evalulation a little cleaner.

This commit introduces a minor constness break, however it's alright.