libtip: split out RgbaImage into a separate file
ditto. Also makes RgbaImage have some useful stuff, instead of
just being a vector and image size pair.

TipImage and RgbaImage now have Clear() methods which clear out
any allocated data, which should mean that memory usage goes down
as the TIP file is exported. Not that memory usage is that big
of an issue, but it's nice to deal with now.