Teach libpixel how to write images
Adds a ImageWriter class. An instance of this class can write
out RgbaImages (see SetImage) to specified output formats; PNG
currently being the only one.

Move STB library usage to libpixel. This will be cleaned up (properly)
when libpixel is a separate repository; for now, we're still
using the in-repo copy, which sucks, but I'm too lazy to set
up submodules just to delete them when the move happens.

Probably the last couple commits before libpixel is moved out.